pyproject.toml
[tool.poetry]
name = "ssafy-custom-news-data"
version = "0.1.0"
description = ""
authors = ["jiwoochris <travelandi01@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
feedparser = "^6.0.11"
kafka-python = "^2.0.2"
apache-flink = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
